# apidoc.yaml
swagger: "2.0"
info:
  description: "This is a epiclogue api document."
  version: "1.0.0"
  title: "epiclogue API"
  contact:
    email: "dladsds123@gmail.com"
host: "api.epiclogue.com"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /boards:
    get:
      tags:
      - "boards"
      summary: "get boards for main page"
      description: ""
      parameters:
      - name: "type"
        in: "query"
        description: "projection type"
        required: false
        type: string
        example: "Comic"
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/board_thumb'
        500:
          description: ERROR
          schema:
            $ref: '#/definitions/API_response_error'
      security:
        - access_token: []

    post:
      tags:
      - "boards"
      summary: "post board to server"
      description: ""
      parameters:
      - name: "Body"
        in: "body"
        required: true
        description: "board info body"
        schema:
          type: object
          properties:
            boardTitle:
              type: string
            boardBody:
              type: string
            category:
              type: number
            pub:
              type: number
            language:
              type: string
            boardImg:
              type: array
              items:
                type: number
      responses:
        201:
          description: OK
        500:
          description: ERROR
          schema:
            $ref: '#/definitions/API_response_error'
      security:
        - access_token: []
...
  /boards/{boardUid}:
    get:
      tags:
      - "boards"
      summary: "get board info for view"
      description: ""
      parameters:
      - name: "boardUid"
        in: "path"
        description: "ID of board"
        required: true
        type: string
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              result:
                type: string
                example: "ok"
              data:
                type: object
                $ref: '#/definitions/board_view'
        500:
          description: ERROR
          schema:
            $ref: '#/definitions/API_response_error'
      security:
        - access_token: []
    
    delete:
      tags:
      - "boards"
      summary: "delete board"
      description: ""
      parameters:
      - name: "boardUid"
        in: "path"
        description: "ID of board"
        required: true
        type: string
      responses:
        200:
          description: OK
        400:
          description: FAIL
        500:
          description: ERROR
          schema:
            $ref: '#/definitions/API_response_error'
      security:
        - access_token: []
...
securityDefinitions:
  access_token:
    description: "insert your accessToken in login response."
    type: apiKey
    name: "x-access-token"
    in: "header"
definitions:
  board_thumb:
    type: "object"
    properties:
      _id: 
        type: string
        example: '1a4c6ea830e14ef67'
      boardTitle:
        type: string
        example: 'latata'
      boardImg:
        type: array
        example: ["link1", "link2"]
      category:
        type: number
        example: 0
      pub: 
        type: number
        example: 1 
      writer:
        type: object
        $ref: '#/definitions/Writer'
      liked:
        type: boolean
        example: true
      bookmarked:
        type: boolean
        example: false
  board_view:
    type: "object"
    required: 
    - "boardImg"
    - "category"
    - "pub"
    properties:
      boardImg:
        type: array
        example: ["link1", "link2"]
      boardTitle:
        type: string
        example: 'WA!'
      boardBody:
        type: string
        example: "SANS!"
      category:
        type: string
        example: "comic"
      pub:
        type: number
        example: 1
      writeDate:
        type: string
        format: date-time
        default: "2018-03-20T09:12:28Z"
      language:
        type: string
        default: "korean"
      heartCount:
        type: number
        default: 0
      feedbackCount:
        type: number
        default: 0
      bookmarkCount:
        type: number
        default: 5
      reactCount:
        type: number
        default: 10
      originUserId:
        type: string
      originBoardId:
        type: string
      allowSecondaryCreation:
        type: string
      edited:
        type: boolean
        default: false
      bookmarked:
        type: boolean
        default: false
      liked:
        type: boolean
        default: false
      writer:
        type: Object
        $ref: '#/definitions/Writer'
      feedbacks:
        type: array
        $ref: '#/definitions/Feedback'
  Writer:
    type: object
    properties:
      _id:
        type: string
        example: '867cab2512312aed1f'
      nickname:
        type: string
        example: 'swagger_awe'
      screenId:
        type: string
        example: 'hippo'
      profile:
        type: string
        example: 'a_link'
      following:
        type: boolean
        example: true
  Feedback:
    type: "object"
    required: 
    - "userId"
    - "boardId"
    - "feedbackBody"
    properties:
      writer:
        type: object
        $ref: '#/definitions/Writer'
      feedbackBody:
        type: string
      writeDate:
        type: string
        format: date-time
        default: "2018-03-20T09:12:28Z"
      edited:
        type: boolean
        default: false
      childCount:
        type: number
        default: 0
      heartCount:
        type: number
        default: 0
      liked:
        type: boolean
        default: true
...
  API_response_error:
    type: "object"
    properties:
      result:
        type: string
        example: "error"
      message:
        type: string